SharingScannerVisitor deleted, it1s functionaliti is moved to SharingVisitor.
